TreeView useReducer WIP
